Fix: Refactor ProductExplorer component with compatible syntax
- Simplify ProductExplorer component without $bindable() runes
- Use standard export let for props (selectedProduct)
- Keep viewProductTree() logic in +page.svelte
- Pass selectedProduct as prop to ProductExplorer
- Remove complex binding and method calls
- Fix svelte.config.js to use adapter-auto instead of adapter-node

Changes:
- ProductExplorer: Simple component with export let selectedProduct
- +page.svelte: Keep API call logic, pass data to component
- Compatible with both Svelte 4 and Svelte 5
- Cleaner separation: logic in page, display in component

The component now only handles display, not data fetching.
This is simpler and more maintainable.
